Struct elastic_codegen::gen::url_params::UrlParamBuilder

pub struct UrlParamBuilder { /* fields omitted */ }

Builder for request url parameters enum.

The output of this structure is an enum that only accepts valid parameter combinations, based on what's given in the paths for an endpoint.

Methods

impl UrlParamBuilder

pub fn new(name: &str) -> Self

pub fn name(&self) -> &Ident

pub fn with_param_set(self, params: Vec<&str>) -> Self

Add a param set to this builder

pub fn build(self) -> (Item, Ty)

Build this enum and return an AST for its declaration and an AST for its type.